requirements.txt
requests
numpy
opencv-python
python_dotenv